refactor: rename project item to just project
