+++ src/app/resume-page/projects-section/get-project-items.ts
@@ -1,9 +1,9 @@
 import { inject, InjectionToken } from '@angular/core'
-import { ProjectItem } from '../data/project-item'
+import { Project } from '../data/project'
 import { ADAPT_JSON_RESUME_PROJECT } from './adapt-json-resume-project'
 import { JSON_RESUME_PROJECTS } from './json-resume-projects'
 
-export type GetProjectItems = () => readonly ProjectItem[]
+export type GetProjectItems = () => readonly Project[]
 
 export const GET_PROJECT_ITEMS = new InjectionToken<GetProjectItems>(
   /* istanbul ignore next */
