+++ src/app/resume-page/projects-section/project-item/project-item-to-contents.ts
@@ -1,11 +1,11 @@
 import { ChippedContent } from '../../chipped-content/chipped-content'
-import { ProjectItem } from '../../data/project-item'
+import { Project } from '../../data/project'
 import { TextContentComponent } from '../../chipped-content/text-content/text-content.component'
 import { ProjectItemTechnologiesComponent } from './project-item-technologies/project-item-technologies.component'
 import { isNotUndefined } from '@/common/is-not-undefined'
 
 export const projectItemToContents: (
-  item: ProjectItem,
+  item: Project,
 ) => readonly ChippedContent[] = (item) =>
   [
     item.description
